requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
numpy>=1.24.0

# PDF parsing (install at least one)
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# DOCX parsing
python-docx>=1.1.0

# Semantic search (strongly recommended — enables hybrid + RAG)
sentence-transformers>=2.7.0

# LLM providers (set corresponding env var)
anthropic>=0.25.0    # for Claude — set ANTHROPIC_API_KEY
# openai>=1.30.0     # for GPT-4o  — set OPENAI_API_KEY
